@@ -2,7 +2,7 @@ apply plugin: 'com.android.library'
 apply plugin: 'com.novoda.bintray-release'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     buildToolsVersion "27.0.3"
 
     defaultConfig {
@@ -19,6 +19,7 @@ android {
 }
 
 publish {
+    userOrg = 'egeniq'
     groupId = 'com.github.stfalcon'
     artifactId = 'chatkit'
     publishVersion = '0.3.3'
@@ -36,7 +37,7 @@ ext {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation "com.android.support:appcompat-v7:$supportVersion"
-    implementation "com.android.support:design:$supportVersion"
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'com.google.android.material:material:1.0.0'
     implementation "com.google.android:flexbox:$flexboxVersion"
 }
